# app/api_doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/upload/add_student_face_from_url": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Student Face From Url",
        "description": "Adds a face to the student's row in the student collection in the database. This is going to be one of the base faces of the student, from which the model trains.\n\n:return: URL of the uploaded image.",
        "operationId": "add_student_face_from_url_upload_add_student_face_from_url_post",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          },
          {
            "name": "face_image_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Face Image Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddFaceResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/add_student_face": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Student Face",
        "description": "Uploads face image to firebase. This is going to be one of the base faces of the student, from which the model trains.\n:return: URL of the uploaded image.",
        "operationId": "add_student_face_upload_add_student_face_post",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_student_face_upload_add_student_face_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddFaceResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/add_class_photo_from_url": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Class Photo From Url",
        "description": "Adds class photo to firebase. This is ideally from PI. Non ideally from teachers phone. \n:return: URL of the uploaded image.",
        "operationId": "add_class_photo_from_url_upload_add_class_photo_from_url_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time"
            }
          },
          {
            "name": "class_photo_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Photo Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddClassPhotoResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/add_class_photo": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Class Photo",
        "description": "Adds class photo to firebase. This is ideally from PI. Non ideally from teachers phone.\n:return: URL of the uploaded image.",
        "operationId": "add_class_photo_upload_add_class_photo_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_class_photo_upload_add_class_photo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddClassPhotoResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/add_attendance": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Attendance",
        "description": "Adds attendance to the database. This is information from the teachers' app from the teacher.\n:param attModel: Attendance model that contains all the necessary information\n:return: URL of the uploaded image.",
        "operationId": "add_attendance_upload_add_attendance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/add_face_encoding": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Add Face Encoding",
        "description": "Adds face encoding to the database. This is done from the server, but a function is written here nonetheless to atomise the process. They are pickle files. \n:param face_encoding_model: Face encoding model that contains all the necessary information\n:return: URL of the uploaded image.",
        "operationId": "add_face_encoding_upload_add_face_encoding_post",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "number_of_faces",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number Of Faces"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_face_encoding_upload_add_face_encoding_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceEncodingModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/update_face_encoding": {
      "post": {
        "tags": ["Upload Images or Attendance Info from App/Website/Pi"],
        "summary": "Update Face Encoding",
        "description": "Updates face encoding in the database. This should also be done from the server, but we cant overwrite or update files in firebase or s3, so we delete the previous file, and we upload the new file, also changing the url in the database, specifically in the student collection. \n:param face_encoding_model: Face encoding model that contains all the necessary information\n:return: URL of the uploaded image.",
        "operationId": "update_face_encoding_upload_update_face_encoding_post",
        "parameters": [
          {
            "name": "old_encoding_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Old Encoding Url"
            }
          },
          {
            "name": "student_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "number_of_faces",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number Of Faces"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_face_encoding_upload_update_face_encoding_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceEncodingResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/student/test": {
      "get": {
        "tags": ["Students"],
        "summary": "Test route",
        "operationId": "index_student_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/student/add_student": {
      "post": {
        "tags": ["Students"],
        "summary": "Add a student",
        "operationId": "add_student_student_add_student_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/face_rec/test": {
      "get": {
        "tags": ["Face Recognition"],
        "summary": "Test route",
        "operationId": "index_face_rec_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/panels/test": {
      "get": {
        "tags": ["Panels, Schools and Specializations"],
        "summary": "Test route",
        "operationId": "index_panels_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/college/test": {
      "get": {
        "tags": ["Rooms and Buildings"],
        "summary": "Test route",
        "operationId": "index_college_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/subjects/test": {
      "get": {
        "tags": ["Subjects and Semesters"],
        "summary": "Test route",
        "operationId": "index_subjects_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/teachers/test": {
      "get": {
        "tags": ["Teachers"],
        "summary": "Test route",
        "operationId": "index_teachers_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/classes/test": {
      "get": {
        "tags": ["Classes and Class Images"],
        "summary": "Test route",
        "operationId": "index_classes_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/test": {
      "get": {
        "tags": ["test"],
        "summary": "Test route",
        "operationId": "index_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddClassPhotoResponseModel": {
        "properties": {
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "format": "time",
            "title": "Time"
          },
          "class_photo_url": {
            "type": "string",
            "title": "Class Photo Url"
          }
        },
        "type": "object",
        "required": ["room_id", "date", "time", "class_photo_url"],
        "title": "AddClassPhotoResponseModel"
      },
      "AddFaceResponseModel": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "face_image_url": {
            "type": "string",
            "title": "Face Image Url"
          }
        },
        "type": "object",
        "required": ["student_id", "face_image_url"],
        "title": "AddFaceResponseModel"
      },
      "AttendanceModel": {
        "properties": {
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "teacher_id": {
            "type": "string",
            "title": "Teacher Id"
          },
          "panel_id": {
            "type": "string",
            "title": "Panel Id"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "subject_id",
          "teacher_id",
          "panel_id",
          "start_time",
          "end_time"
        ],
        "title": "AttendanceModel"
      },
      "Body_add_class_photo_upload_add_class_photo_post": {
        "properties": {
          "class_photo": {
            "type": "string",
            "format": "binary",
            "title": "Class Photo"
          }
        },
        "type": "object",
        "required": ["class_photo"],
        "title": "Body_add_class_photo_upload_add_class_photo_post"
      },
      "Body_add_face_encoding_upload_add_face_encoding_post": {
        "properties": {
          "face_encoding": {
            "type": "string",
            "format": "binary",
            "title": "Face Encoding"
          }
        },
        "type": "object",
        "required": ["face_encoding"],
        "title": "Body_add_face_encoding_upload_add_face_encoding_post"
      },
      "Body_add_student_face_upload_add_student_face_post": {
        "properties": {
          "face_image": {
            "type": "string",
            "format": "binary",
            "title": "Face Image"
          }
        },
        "type": "object",
        "required": ["face_image"],
        "title": "Body_add_student_face_upload_add_student_face_post"
      },
      "Body_update_face_encoding_upload_update_face_encoding_post": {
        "properties": {
          "face_encoding": {
            "type": "string",
            "format": "binary",
            "title": "Face Encoding"
          }
        },
        "type": "object",
        "required": ["face_encoding"],
        "title": "Body_update_face_encoding_upload_update_face_encoding_post"
      },
      "FaceEncodingModel": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "number_of_faces": {
            "type": "integer",
            "title": "Number Of Faces"
          },
          "encoding": {
            "type": "string",
            "format": "binary",
            "title": "Encoding"
          }
        },
        "type": "object",
        "required": ["student_id", "number_of_faces", "encoding"],
        "title": "FaceEncodingModel"
      },
      "FaceEncodingResponseModel": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "number_of_faces": {
            "type": "integer",
            "title": "Number Of Faces"
          },
          "encoding_url": {
            "type": "string",
            "title": "Encoding Url"
          }
        },
        "type": "object",
        "required": ["student_id", "number_of_faces", "encoding_url"],
        "title": "FaceEncodingResponseModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "StudentModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prn": {
            "type": "string",
            "title": "Prn"
          },
          "panel": {
            "type": "string",
            "title": "Panel"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "face_encoding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Encoding Id"
          },
          "panel_roll_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Roll Number"
          },
          "faces": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faces"
          }
        },
        "type": "object",
        "required": ["name", "prn", "panel", "email"],
        "title": "StudentModel"
      },
      "StudentResponseModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prn": {
            "type": "string",
            "title": "Prn"
          },
          "panel": {
            "type": "string",
            "title": "Panel"
          },
          "panel_roll_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Roll Number"
          }
        },
        "type": "object",
        "required": ["id", "name", "prn", "panel"],
        "title": "StudentResponseModel"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}